Add roles and labels to Luna nav
Tested using the ChromeVox extension.

I added the aria-label to each version of the header  nav in order to distinguish
them from the footer nav. The screen reader will now include the aria-label
in its description of the nav. That choice was based on [this MDN
resource](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Navigation_Role)
The footer nav did not need an extra aria-label because it is nested
inside of the footer element, and the screen reader already includes the
word "footer"-- the aria-label would have also said "footer" and been
redundant.

For the screen reader, the role="navigation" attr had to be present in
order for the aria-label to be included in the spoken description.

[PT Story](https://www.pivotaltracker.com/story/show/169550274)