Add contentinfo role to footer
This addresses a bug in Apple VoiceOver where the footer content isn't
announced. Assigning it a role of "contentinfo" allows VoiceOver to
recognize it as a footer.

Previously, I had left out an aria-label for the footer nav to avoid
screen reader repetition of the word "footer," but "contentinfo" seems
to work well with aria-label="Footer".

[PT Story](https://www.pivotaltracker.com/story/show/169649155)
[Resource](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/footer#Accessibility_concerns)