Convert remove item link to button
Converting from a link to a button allows users to interact with buttons
with either the space key or enter key-- ie we enable default button
behaviors. This element is a button because it triggers an action and
does not navigate to another page or view.

Fixes https://www.pivotaltracker.com/story/show/169827557