Remove bottom category nav
Because it was a duplicate of the category nav at the top of the product
grid, it violated the rule that landmarks must be unique. It could be
challenging or frustrating for a user relying on accessibility tools to
figure out if there was a difference between the 2 navs. The project
team decided it was okay to remove the second nav.

Fixes https://www.pivotaltracker.com/story/show/170060356
Related to
https://www.pivotaltracker.com/n/projects/1902125/stories/169828506